+++ contracts/LinearReverseDutchAuction.sol
@@ -26,6 +26,7 @@ contract LinearReverseDutchAuction is Auction {
         uint256 startingPrice;
         uint256 availableFunds;
         uint256 reservedPrice;
+        uint256 settle_price;
         address winner;
         uint256 deadline;
         uint256 duration;
@@ -75,6 +76,7 @@ contract LinearReverseDutchAuction is Auction {
             startingPrice: startingPrice,
             availableFunds: 0,
             reservedPrice: reservedPrice,
+            settle_price: reservedPrice,
             winner: msg.sender,
             deadline: deadline,
             duration: duration,
@@ -85,7 +87,7 @@ contract LinearReverseDutchAuction is Auction {
 
     function getCurrentPrice(uint256 auctionId) public view validAuctionId(auctionId) returns (uint256) {
         AuctionData storage auction = auctions[auctionId];
-        if(block.timestamp >= auction.deadline) return 0;
+        if(block.timestamp >= auction.deadline) return auction.settle_price;
         require(!auction.isClaimed, 'Auction has ended');
         // price(t) = startingPrice - (((startingPrice - reservedPrice) * (timeElapsed)) / duration)
         return auction.startingPrice - (((auction.startingPrice - auction.reservedPrice) * (block.timestamp - (auction.deadline - auction.duration))) / auction.duration);
@@ -94,12 +96,13 @@ contract LinearReverseDutchAuction is Auction {
     //placeBid function is not required in this auction type as the price is determined by the auctioneer and the auctioneer can withdraw the item on placing the bid only
     function withdrawItem(uint256 auctionId) external validAuctionId(auctionId) {
         AuctionData storage auction = auctions[auctionId];
-        require(block.timestamp < auction.deadline || auction.winner==auction.auctioneer, 'Auction has ended');
+        require((block.timestamp < auction.deadline && msg.sender!=auction.auctioneer) || (auction.winner==auction.auctioneer && block.timestamp >= auction.deadline && msg.sender == auction.auctioneer), 'Not allowed to withdraw item');
         require(!auction.isClaimed, 'Auction has been settled');
         uint256 currentPrice = getCurrentPrice(auctionId);
         auction.winner = msg.sender;
         auction.availableFunds = currentPrice;
         auction.isClaimed = true;
+        auction.settle_price = currentPrice;
         if(auction.auctioneer != auction.winner) receiveFunds(false, auction.biddingToken, msg.sender, currentPrice);
         sendFunds(auction.auctionType == AuctionType.NFT, auction.auctionedToken, msg.sender, auction.auctionedTokenIdOrAmount);
         emit itemWithdrawn(auctionId, msg.sender, auction.auctionedToken, auction.auctionedTokenIdOrAmount);
